Finished First Iteration.
You can now attach this behavior to any model, and when saving add field called 'draft' to your data array as you are saving, and instead of saving to the existing record it will save to the drafts table instead.

This is good for saving versions without needing to have it be the live version of a record.  Examples might be a CMS when you update a webpage but don't want the world to see it yet.  Or you update a blog post and don't want to update the live version just yet.